tankPython/requirements.txt
Flask==1.1.2
websocket-client==0.57.0